requirements.txt
django
django-mapbox-location-field